move info session from `/api/infoSession/user` to `/api.signups/info`
